+++ ts/helpers/index.ts
@@ -1,0 +1,5 @@
+import { getAttributeByPath } from "./attributesHandler";
+import { CheckValidObject as checkPath } from "./checkobj";
+import ZipHandler from "./ziphandler";
+
+export { getAttributeByPath, checkPath, ZipHandler };
