+++ ts/parsers/lineparser.ts
@@ -1,4 +1,4 @@
-import { CheckValidObject as checkPath } from "../helpers/checkobj";
+import { checkPath } from "../helpers";
 import { PowerpointElement, BorderType } from "airppt-models-plus/pptelement";
 
 /**
